+++ Main.Documents.cs
@@ -388,6 +388,7 @@ namespace DDF___Program_Language_Editor
             public TabPage Tab { get; }
             public Stack<Tuple<int, int>> SelectionHistory { get; } = new Stack<Tuple<int, int>>();
             public List<DdfTextRange> MultiSelections { get; } = new List<DdfTextRange>();
+            public int? LastEditPosition { get; set; }
         }
     }
 }